+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 ==============
 Changelog
 ==============
+2021-02-12 (version 1.17):
+    * move to next/previous observation with arrow right/left
+    * move to next/previous observation window with CTRL + arrow right/left or A/D
+    * move to last/first observation with End/Pos1 or ALT + A/D
+    * select next/previous vector feature with arrow downs/up or S/W
+    * show next/previous map view with PageDown/PageUp or ALT + S/W
+    * modified observation slider, slider shows range of visible dates
+
 2021-02-02 (version 1.16):
     * fixed smaller issues
     * forward / backward button to move in time now shifts by number of opened observation dates/maps